put serialize balance and its unit tests into correct place.
