+++ driver/src/models.rs
@@ -3,6 +3,9 @@ use serde_derive::{Deserialize, Serialize};
 use sha2::{Digest, Sha256};
 use web3::types::H256;
 use crate::price_finding::Solution;
+use crate::price_finding::linear_optimization_price_finder::{account_id, token_id};
+use std::collections::HashMap;
+use serde_json::json;
 
 pub const TOKENS: u8 = 30;
 pub const DB_NAME: &str = "dfusion2";
@@ -43,8 +46,21 @@ impl State {
     pub fn decrement_balance(&mut self, token_id: u8, account_id: u16, amount: u128) {
         self.balances[State::balance_index(token_id, account_id)] -= amount;
     }
-    pub fn read_balances(&self) -> Vec<u128> {
-        self.balances.clone()
+    pub fn serialize_balances(&self, num_tokens: u8) -> serde_json::Value {
+        assert_eq!(
+            self.balances.len() % num_tokens as usize, 0,
+            "Balance vector cannot be split into equal accounts"
+        );
+        let mut accounts: HashMap<String, HashMap<String, String>> = HashMap::new();
+        let mut current_account = 0;
+        for account_balances in self.balances.chunks(num_tokens as usize) {
+            accounts.insert(account_id(current_account), (0..num_tokens)
+                .map(token_id)
+                .zip(account_balances.iter().map(|b| b.to_string()))
+                .collect());
+            current_account += 1;
+        }
+        json!(accounts)
     }
 
     pub fn update_balances(&mut self, orders: &[Order], solution: &Solution) {
@@ -210,7 +226,7 @@ pub mod tests {
   use std::str::FromStr;
 
     #[test]
-    fn computes_updated_balance_on_example_with_equal_buy_and_sell(){
+    fn test_update_balances(){
         let mut state = State::new(
             "test".to_string(),
             0,
@@ -245,6 +261,40 @@ pub mod tests {
         assert_eq!(state.read_balance(1, 0), 99);
         assert_eq!(state.read_balance(0, 1), 99);
         assert_eq!(state.read_balance(1, 1), 101);
+    }
+
+    #[test]
+    fn test_serialize_balances() {
+        let state = State::new(
+            "test".to_string(),
+            0,
+            vec![100, 200, 300, 400, 500, 600]
+        );
+        let result = state.serialize_balances(3);
+        let expected = json!({
+            "account0": {
+                "token0": "100",
+                "token1": "200",
+                "token2": "300",
+            },
+            "account1": {
+                "token0": "400",
+                "token1": "500",
+                "token2": "600",
+            }
+        });
+        assert_eq!(result, expected)
+    }
+
+    #[test]
+    #[should_panic]
+    fn test_serialize_balances_with_bad_balance_length() {
+        let state = State::new(
+            "test".to_string(),
+            0,
+            vec![100, 200]
+        );
+        state.serialize_balances( 3);
     }
 
   #[test]
